Big reset
I need to focus on the actual prose and not nonsense worldbuilding
"organization." From here on, at least for now, I'm not going to worry
about how this will render on my website or how I ought to have a unified
system for making graphs, diagrams, etc. The only actually important
part of writing is raw prose - it's much more fun to write too.